Refactor file structure & build steps
