+++ src/index.ts
@@ -1,4 +1,4 @@
-import { JestChrome } from '../jest-chrome'
+import { JestChrome } from './jest-chrome'
 import { createHandler } from './createHandler'
 
 export const chrome = new Proxy<JestChrome>(
